# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movies & Reviews API",
    "version": "1.0.0",
    "description": "API with full CRUD for Movies and Reviews with GitHub OAuth login"
  },
  "servers": [
    {
      "url": "https://project3-zo0c.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/auth/github": {
      "get": {
        "summary": "Login with GitHub",
        "description": "Redirects the user to GitHub for OAuth login.",
        "tags": [
          "Auth"
        ],
        "responses": {
          "302": {
            "description": "Redirects to GitHub OAuth"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "summary": "GitHub OAuth callback",
        "description": "Handles the GitHub OAuth callback and redirects to Swagger UI on success.",
        "tags": [
          "Auth"
        ],
        "responses": {
          "302": {
            "description": "Redirects to Swagger UI on success or home on failure"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "Logout the current user",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "500": {
            "description": "Logout failed"
          }
        }
      }
    },
    "/api/movies": {
      "get": {
        "summary": "Get all movies",
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "Returns an array of movies"
          }
        }
      },
      "post": {
        "summary": "Create a new movie",
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "director",
                  "rating"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "director": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "releaseYear": {
                    "type": "number"
                  },
                  "rating": {
                    "type": "number"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "cast": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movie created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/movies/{id}": {
      "put": {
        "summary": "Update a movie by ID",
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The movie ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "director": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "releaseYear": {
                    "type": "number"
                  },
                  "rating": {
                    "type": "number"
                  },
                  "duration": {
                    "type": "number"
                  },
                  "cast": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Movie updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Movie not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a movie by ID",
        "tags": [
          "Movies"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The movie ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Movie not found"
          }
        }
      }
    },
    "/api/reviews/{movieId}": {
      "get": {
        "summary": "Get all reviews for a movie",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "movieId",
            "required": true,
            "description": "The ID of the movie",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of reviews"
          }
        }
      }
    },
    "/api/reviews": {
      "post": {
        "summary": "Create a new review",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "movieId",
                  "user",
                  "rating",
                  "comment"
                ],
                "properties": {
                  "movieId": {
                    "type": "string",
                    "description": "ID of the movie being reviewed"
                  },
                  "user": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 10
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/reviews/{id}": {
      "put": {
        "summary": "Update a review by ID",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the review",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 10
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a review by ID",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the review",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Review not found"
          }
        }
      }
    }
  },
  "tags": []
}